Move the C++ `Trap` and related types to its own header
